Make ProgressFileReader more robust
Progress file dir is deleted if it exists only (it may not exist if start() is
not called).

Rely on value() methods to get a String representation of the value.

+++ scheduler/scheduler-node/src/main/java/org/ow2/proactive/scheduler/task/ProgressFileReader.java
@@ -57,7 +57,7 @@ public class ProgressFileReader {
 
     public boolean start(File workingDir, TaskId taskId) {
         String progressFileName =
-                "job-" + taskId.getJobId() + "-task-" + taskId.toString()
+                "job-" + taskId.getJobId().value() + "-task-" + taskId.value()
                         + "-" + UUID.randomUUID() + ".progress";
 
         return start(workingDir, progressFileName);
@@ -134,6 +134,11 @@ public class ProgressFileReader {
     }
 
     private void removeProgressFileDir() {
+        if (progressFileDir == null
+                || !Files.exists(progressFileDir)) {
+            return;
+        }
+
         try {
             FileUtils.deleteRecursively(progressFileDir);
         } catch (IOException e) {
